+++ test/ServerStatManTest.cc
@@ -127,14 +127,10 @@ void ServerStatManTest::testLoad()
   CPPUNIT_ASSERT(localhost_http);
   CPPUNIT_ASSERT_EQUAL(std::string("localhost"), localhost_http->getHostname());
   CPPUNIT_ASSERT_EQUAL(std::string("http"), localhost_http->getProtocol());
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(25000),
-                       localhost_http->getDownloadSpeed());
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(101),
-                       localhost_http->getSingleConnectionAvgSpeed());
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(102),
-                       localhost_http->getMultiConnectionAvgSpeed());
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(6),
-                       localhost_http->getCounter());
+  CPPUNIT_ASSERT_EQUAL(25000, localhost_http->getDownloadSpeed());
+  CPPUNIT_ASSERT_EQUAL(101, localhost_http->getSingleConnectionAvgSpeed());
+  CPPUNIT_ASSERT_EQUAL(102, localhost_http->getMultiConnectionAvgSpeed());
+  CPPUNIT_ASSERT_EQUAL(6, localhost_http->getCounter());
   CPPUNIT_ASSERT_EQUAL(static_cast<time_t>(1210000000),
                        localhost_http->getLastUpdated().getTime());
   CPPUNIT_ASSERT_EQUAL(ServerStat::OK, localhost_http->getStatus());
